- moving tests down into directory test  - tests for align  - update align to satisfy all the tests

+++ toolframe.py
@@ -208,7 +208,7 @@ def ez_launch(main = None, cleanup=None, test=None):
         return
     sname = sys.argv[0]
     pname = re.sub('.py$', '', sname)
-    if sname.endswith('.py') and not os.path.exists(pname):
+    if sname.endswith('.py') and not os.path.exists(pname) and '-L' in sys.argv:
         print("creating symlink: %s -> %s" % (pname, sname))
         os.symlink(sname, pname)
     elif sys._getframe(1).f_code.co_name in ['?', '<module>']:
@@ -219,7 +219,7 @@ def ez_launch(main = None, cleanup=None, test=None):
             if test == None:
                 unittest.main()
             else:
-                if not testhelp.main(sys.argv,test) and cleanup != None:
+                if not testhelp.main(sys.argv) and cleanup != None:
                     cleanup()
         elif main == None:
             raise StandardError("Pass your main routine to ez_launch")
